Fixed id, secret, and spotify api
Client ID and Client Secret are now stored in another file. The Spotify API is connected but the user is hard coded to 'theadoxbox' at the moment. The API will get all public playlists from this user, print the album name, print all the tracks, and print the image as well